requirements.txt
django

django-environ
django-cleanup
psycopg2-binary
django-tailwind
django-browser-reload

django-storages
django-cors-headers
django-ratelimit
dj-database-url
djangorestframework-simplejwt

pytz

django-phonenumber-field[phonenumberslite]

google-cloud-storage
google-api-python-client

whitenoise

pillow

gunicorn

boto3
openpyxl
